Move slack callbacks into callables subdirectory.
