+++ ea_airflow_util/__init__.py
@@ -5,5 +5,5 @@ from ea_airflow_util.dags.s3_to_snowflake_dag import S3ToSnowflakeDag
 from ea_airflow_util.dags.dbt_snapshot_dag import DbtSnapshotDag
 from ea_airflow_util.dags.sftp_to_snowflake_dag import SFTPToSnowflakeDag
 
-import ea_airflow_util.dags.dag_util.slack_callbacks as slack_callbacks
-from ea_airflow_util.dags.callables.variable import check_variable, update_variable
+import ea_airflow_util.callables.slack as slack_callbacks
+from ea_airflow_util.callables.variable import check_variable, update_variable
