+++ ea_airflow_util/dags/dbt_snapshot_dag.py
@@ -1,7 +1,7 @@
 from functools import partial
 from typing import Optional
 
-import ea_airflow_util.dags.dag_util.slack_callbacks as slack_callbacks
+from ea_airflow_util.callables import slack
 
 from airflow import DAG
 from airflow_dbt.operators.dbt_operator import DbtSnapshotOperator
@@ -46,7 +46,7 @@ class DbtSnapshotDag():
         """
         # If a Slack connection has been defined, add the failure callback to the default_args.
         if self.slack_conn_id:
-            slack_failure_callback = partial(slack_callbacks.slack_alert_failure, http_conn_id=self.slack_conn_id)
+            slack_failure_callback = partial(slack.slack_alert_failure, http_conn_id=self.slack_conn_id)
             default_args['on_failure_callback'] = slack_failure_callback
 
         return DAG(
