+++ ea_airflow_util/dags/s3_to_snowflake_dag.py
@@ -10,7 +10,7 @@ from airflow.providers.amazon.aws.operators.s3 import S3ListOperator
 from airflow.providers.snowflake.hooks.snowflake import SnowflakeHook
 from airflow.utils.helpers import chain
 
-import ea_airflow_util.dags.dag_util.slack_callbacks as slack_callbacks
+from ea_airflow_util.callables import slack
 from .operators.loop_s3_file_transform_operator import LoopS3FileTransformOperator
 from .dag_util.xcom_util import xcom_pull_template
 
@@ -86,11 +86,11 @@ class S3ToSnowflakeDag():
         """
         # If a Slack connection has been defined, add the failure callback to the default_args.
         if self.slack_conn_id:
-            slack_failure_callback = partial(slack_callbacks.slack_alert_failure, http_conn_id=self.slack_conn_id)
+            slack_failure_callback = partial(slack.slack_alert_failure, http_conn_id=self.slack_conn_id)
             default_args['on_failure_callback'] = slack_failure_callback
 
             # Define an SLA-miss callback as well.
-            slack_sla_miss_callback = partial(slack_callbacks.slack_alert_sla_miss, http_conn_id=self.slack_conn_id)
+            slack_sla_miss_callback = partial(slack.slack_alert_sla_miss, http_conn_id=self.slack_conn_id)
         else:
             slack_sla_miss_callback = None
